Release instance lock if unpacking did not complete
Summary:
There are two locks when initializing SoLoader:
- The first one is the blocking lock, whose purpose is to make sure only one process is unpacking libraries from the APK at a time.
- The second one is the instance lock, whose purpose is to make sure that a process doesn't unpack libraries when there is another process loading libraries from the unpacked directory.

The second lock was introduced in  D29935532 (https://github.com/facebook/SoLoader/commit/efdd618edc867ebbe7195eec0c76a1b3fe3684f2) to try to address a longstanding crash in which corrupt libraries are loaded. The way the instance lock works, is that the first process to get this lock will try to unpack libraries, and it will then hold this lock until the process ends. No other process will be able to unpack libraries while this process holds the lock, to avoid that process from trying to load a library while another process is overwriting the library. The problem with this implementation is that if the first process grabs the lock, and then unpacking fails for some reason, it will still hold the lock and not allow another process to retry unpacking. This diff fixes that issue by releasing the instance lock if unpacking did not finish successfully.

Reviewed By: simpleton

Differential Revision: D43146532

fbshipit-source-id: bfa31cda7d8df6d93eef8103eda6954231530389