Wait for UnpackingSoSources to finish unpacking after failure to load library
Summary:
SoLoader sees a variety of crashes that we believe stem from attempting to use stale libraries.

~~We believe one such scenario can occur during app updates, if various classes are using a stale/incorrect `ApplicationContext`, and not regenerating as they should.~~

~~After we fail to load a library, `ApplicationSoSource` regenerates itself if it notices the `nativeLibDir` has changed. ~~

~~This diff causes `SoLoader` to regenerate any `UnpackingSoSource` when that happens.~~

~~Idea being, if `nativeLibDir` changed -> APK has changed ->  APK has new SoSources we need to unpack.~~

## Edit:
I've changed the approach here to be more conservative.
Rather than recreating UnpackingSoSources, we'll wait for them to unpack.

This in combination with D43677645 should help address cases where we might be using stale due to a stale context.

Differential Revision: D43375578

fbshipit-source-id: 025e2928c8f31610993b361afab2100bad78bdf2