introduce RecoveryStrategy interface
Summary:
The motivation is to introduce an architecture allowing us to exercise a number of recovery mechanism in succession, until one that works is found. For exact contract see RecoveryStrategyFactory.

To showcase the new abstractions this diff migrates `loadLibraryBySoName` to use the new framework (waiting for unpacking, refreshing context, checking whether the base apk exists).

Reviewed By: adicatana

Differential Revision: D46146373

fbshipit-source-id: 0813196cdac1e58bf3f83f8853ce456a239c5a4b