Always extract from ApkSoSource for /data/app failures
Summary:
After D48070086, we're currently not re-unpacking if the /data/app libraries are corrupted, but has the expected number of bytes. In that diff we removed an important check in ApkSoSource line 79, which is checking the name of the crashing library against the one that we're checking for extraction.

This diff aims to solve that issue with the following logic:
* if there's a /data/app crash, then kick off unconditional unpacking of whatever is zipped under base.apk!/lib/arm64

Reviewed By: adicatana

Differential Revision: D49144930

fbshipit-source-id: 43b8a68e79d110fc271a1ecd3f2b5bfa8684c42c