Include short name in systrace block name
Summary:
We log the soname in the systrace blocks, which in prod resolves to the name of the merged library. This leads to traces like https://fburl.com/tracery/5sr15q1q, where we know which merged lib we are loading, but we don't have a way to identify the JNI entry point which triggered the load.

There are many cases (like the one described above) where we see a library getting loaded, but can't reproduce locally.
If we log the short name in addition to so name, debugging these issues will become trivial.

Reviewed By: adicatana

Differential Revision: D53679442

fbshipit-source-id: fa92e00d7139def5e33b7707c2b0626c1bef7ead