More granular synchronization when calling JNI_OnLoad on merged libraries
Summary: We're currently blocking on a lock corresponding to the merged library name, even though Jni_OnLoads can be run in parallel. That is, if we have three libraries A, B, C merged into libmerged.so, as soon as libmerged has been loaded, there's no need to block further when running JNI_OnLoad, iti can (and should) be done in parallel for each library. The only synchronization that needs to be done is at shortName level, e.g. if we have two parallel loadLibrary("A") calls.

Reviewed By: simpleton

Differential Revision: D58954597

fbshipit-source-id: 7ad2eeccbbad9404050cf07c713a78b87b094d5c