Run prepare() for BackupSoSource to cleanup previous state
Summary:
Currently, the BackupSoSource does not clean after itself when a new version of the app is installed. That is because prepare() does not run unless there is a crash.

This diff changes that by running prepare() with special flags whenever initialising SoLoader. With the new flag, we make sure that `deleteSoFiles` runs when necessary, e.g. when deps have changed.

To differentiate between different states, we simply skip writing sos to disk and the deps state (which is empty).

Reviewed By: danjin250

Differential Revision: D60186847

fbshipit-source-id: e029b9c31ac9a677cae5e9c9d80cf8aee79da35d