Allow OSS libraries to provide an ExternalSoMapping during init. (#127)
Summary:
Pull Request resolved: https://github.com/facebook/SoLoader/pull/127

As of now, MergedSoMapping was stabbed to a static empty class, so it was not
possible for libraries or apps in OSS to provide an implementation for it.

I'm extending the SoLoader API to offer a `init` method which accepts a
new interface called `ExternalSoMapping` which is not static, and libraries can
implement.

Reviewed By: adicatana

Differential Revision: D60110088

fbshipit-source-id: 9e47328dc59f27177ab3a221bcbb497bcf2d642b