Remove async option when parsing native_deps file
Summary: The async option does not make a lot of sense anymore. It is the default for all apps using FbSoLoader, whereas the wearables ones do not extract the file to disk anyways. In general, we're moving towards shipping the `native_deps` file uncompressed everywhere, and reading it directly from the APK, synchronously, which is pretty fast.

Differential Revision: D73772666

fbshipit-source-id: 532554fff0f7ce512cc8d5b04350da08cbeb0abe