Load dependencies conditionally
Summary:
Currently Manifest consists of names of libraries. This diff extends it with a per-library bit mask. First bit in the mask tells `DirectSplitSoSource` whether to manually resolve dependencies.

The bit forcing explicit loading of dependencies is set if a library or one of its recursive dependencies references an unknown library (library not present in the split that is not in the allow-list of known system libraries)

Reviewed By: adicatana

Differential Revision: D75373333

fbshipit-source-id: b4e91fd5620f4daa01f9216d69c1514cdaac81d5