build abi config splits
Summary:
# Bundle config
With this diff SoLoader's postprocessor updates BundleConfig to enable UncompressNativeLibraries optimisation with 16kb pages and splits native libs to a separate config split.

# DirectSplitSoSource
`DirectSplitSoSource` is now initialized with a name of a feature. Once manifest is parsed and ABI is known it will choose one of the following (whichever exists, in that order):
* abi config split for the feature
* feature split (if no abi splitting is enabled)
* base.apk (if universal apk)

Reviewed By: bchang7

Differential Revision: D75400595

fbshipit-source-id: 92a19a4dd21c64b3da9a25e4ff89060a8ffca2a2