Move split related logic out of DirectSplitSoSource
Summary: Now that we know the abi to use at construction time, `DirectSplitSoSource` does not need to delay detection of abi split until prepare. This diff adds a constructor of `DirectSplitSoSource` that allow us to explicitly specify name of the split and moves all the code for detecting split names/paths to a helper class `Splits`. `DirectSplitSoSource` is much simpler now.

Reviewed By: adicatana

Differential Revision: D80009440

fbshipit-source-id: a422777f7f138e85133dcedfb3ab1e575d6a1228