Do not register non-asset libs for built-in modules and universal apks
Summary: `SoSourceRegistry.registerApplicationSplit` needs to deal with built-in modules and universal apks with all modules glued into single apk. When we pass base.apk as a feature module to `SoSourceRegistry.registerApplicationSplit` it should not look for non-asset libraries - those are included in base.apk's non asset libs and should not be extracted again.

Reviewed By: bchang7

Differential Revision: D83686228

fbshipit-source-id: 948792ea761241b1f3e9897b8b4d7bf6c9892f62