Use MC param to enable superattic
Summary: Previous diff (D84610647) was reverted because of a primary/secondary cross-reference issue. This is fixed now in this diff, by using reflection to check if a class/method is available or not. This only occurs in fblite because SuperAttic experiment is only available for facebook, not fblite, but FbSoLoader is shared across many apps.

Reviewed By: sb98052, michalgr

Differential Revision: D84917631

fbshipit-source-id: 51227563d95930ea8972b22ff425e5918b70b0c6